refactor(webview): split App.tsx (oh-tab-vfsl)
